add store_local
