+++ finding_leftover/account/templates/signup.html
@@ -1,105 +1,94 @@
 {% extends 'layout.html' %}
 
 {% block content %}
-<!--
-<body>
-    <div class="account" style="width: 30%;">
-    <form method="POST" action="{% url 'signup' %}" enctype="multipart/form-data">
-        {% csrf_token %}
-        <div class="form-group row">
-          <label class="col-sm-2 col-form-label">username</label>
-          <div class="col-sm-10">
-            <input type="text" class="form-control" name="username">
-          </div>
-        </div>
-        <div class="form-group row">
-          <label for="inputPassword3" class="col-sm-2 col-form-label">Password</label>
-          <div class="col-sm-10">
-            <input type="password" class="form-control" id="inputPassword3" name="password">
-          </div>
-        </div>
-        <div class="form-group row">
-            <label for="inputPassword3" class="col-sm-2 col-form-label">confirm</label>
-            <div class="col-sm-10">
-              <input type="password" class="form-control" id="inputPassword3" name="confirm" placeholder="password를 재입력해주세요">
-            </div>
-          </div>
-        
-        <div class="form-group row">
-            <label class="col-sm-2 col-form-label">store name</label>
-            <div class="col-sm-10">
-              <input type="text" class="form-control" name="store_name">
-            </div>
-          </div>
 
-          <div class="form-group row">
-            <label class="col-sm-2 col-form-label">store address</label>
-            <div class="col-sm-10">
-              <input type="text" class="form-control" name="store_address" placeholder="OO시 OO구">
-            </div>
-          </div>
-
-          <div class="form-group row">
-            <label class="col-sm-2 col-form-label">store memo</label>
-            <div class="col-sm-10">
-              <input type="textarea" class="form-control" name="store_memo" placeholder="가게를 한 줄로 소개해주세요">
-            </div>
-          </div>
-
-          <div class="form-group row">
-            <label class="col-sm-2 col-form-label">store image</label>
-            <div class="col-sm-10">
-              <input type="file" name="store_image">
-            </div>
-          </div>
-        <div class="form-group row">
-          <div class="col-sm-10">
-            <button type="submit" class="btn btn-info">가입하기</button>
-          </div>
-        </div>
-      </form>
-
-      
-      <form method="POST" action="{% url 'signup' %}" enctype="multipart/form-data">
-        {% csrf_token %}
-        
-        <p>username</p>
-        <input type="text" name="username">
-        <br>
-        <p>password</p>
-        <input type="text" name="password">
-        <br>
-        <p>confirm</p>
-        <input type="text" name="confirm">
-        <p>store name</p>
-        <input type="text" name="store_name">
-        <br>
-        <p>store address</p>
-        <input type="text" name="store_address">
-        <br>
-        <p>store memo</p>
-        <input type="text" name="store_memo">
-        <br>
-        
-        <br><br>
-        <input type="submit" value="회원가입">
-    </form> 
-    
-</div>
-</body> -->
-
-<!--body는 정서님이 block이전에 선언해놓으셨기때문에 뺐습니다!-->
 <form method="POST" action="{% url 'signup' %}"  enctype="multipart/form-data">
     {% csrf_token %}
     <div class="row">
       <div class="col-6 offset-3">
         <h2><strong>회원 가입</strong></h2><br/>
         {{ signup_form.as_p }}
-        <button type="submit" class="btn btn-info offset-10">가입하기</button>
+        <div class="col">
+          <div class="row mb-1">
+            <input type="text" class="form-control col-3" id="postcode"  placeholder="우편번호"> &nbsp;
+            <input type="button" class="btn btn-outline-info" onclick="execDaumPostcode()" value="우편번호 찾기">
+          </div>
+          <div class="row mb-1">
+          <input type="text" class="form-control " id="address" name="address" placeholder="주소">
+          </div>
+          <div class="row mb-1">
+            <input type="text" id="detailAddress" class="form-control col-6" name="detailAddress" placeholder="상세주소">&nbsp;
+            <input type="text" id="extraAddress" class="form-control col-5" name="extraAddress" placeholder="참고항목">
+          </div>
+        </div>
+        <br/>
+        <input type = hidden id="local" name="local">
+        <button type="submit" class="btn btn-info offset-10 btn-lg">가입하기</button>
       </div>
     </div>
     
 </form>
+
+<!--주소 API(다음우편번호 서비스)추가 -->
+<script src="http://dmaps.daum.net/map_js_init/postcode.v2.js?autoload=false"></script>
+<script>
+    //load함수를 이용하여 core스크립트의 로딩이 완료된 후, 우편번호 서비스를 실행합니다.
+  function execDaumPostcode() {
+    daum.postcode.load(function(){
+        new daum.Postcode({
+            oncomplete: function(data) {
+                // 팝업에서 검색결과 항목을 클릭했을때 실행할 코드
+                var addr = ''; // 주소
+                var extraAddr = ''; // 참고항목
+
+                if (data.userSelectedType === 'R') { // 사용자가 도로명 주소를 선택
+                    addr = data.roadAddress;
+                } else { // 사용자가 지번 주소를 선택(J)
+                    addr = data.jibunAddress;
+                }
+
+                // 사용자가 선택한 주소가 도로명 타입일때 참고항목을 조합한다.
+                if(data.userSelectedType === 'R'){
+                    // 법정동명이 있을 경우 추가한다. (법정리는 제외)
+                    // 법정동의 경우 마지막 문자가 "동/로/가"로 끝난다.
+                    if(data.bname !== '' && /[동|로|가]$/g.test(data.bname)){
+                        extraAddr += data.bname;
+                    }
+                    // 건물명이 있고, 공동주택일 경우 추가
+                    if(data.buildingName !== '' && data.apartment === 'Y'){
+                        extraAddr += (extraAddr !== '' ? ', ' + data.buildingName : data.buildingName);
+                    }
+                    // 표시할 참고항목이 있을 경우, 괄호까지 추가한 최종 문자열을 만든다.
+                    if(extraAddr !== ''){
+                        extraAddr = ' (' + extraAddr + ')';
+                    }
+                    document.getElementById("extraAddress").value = extraAddr;
+                
+                } else {
+                    document.getElementById("extraAddress").value = '';
+                }
+
+                document.getElementById('postcode').value = data.zonecode;
+                document.getElementById("address").value = addr;
+                
+                //custom
+                //시도를 나타내는 변수 : sido
+                //시군구 나타는 변수 : sigungu
+                var sido_list=['서울', '부산', '대구', '인천', '광주', '대전', '울산']
+                if (sido_list.includes(data.sido)){
+                  document.getElementById("local").value = data.sido+"시 "
+                }else{
+                  document.getElementById("local").value = data.sido+" "
+                }
+                document.getElementById("local").value += data.sigungu;
+                  
+                // 커서를 상세주소 필드로 이동
+                document.getElementById("detailAddress").focus();
+            }
+        }).open();
+    });
+  }
+</script>
 
 {% endblock %}
